Extract groupIntoRows
