+++ frontend/src/components/CheckboxGroup/CheckboxGroup.tsx
@@ -2,6 +2,7 @@ import React, { FunctionComponent, useEffect, useState } from 'react';
 import { Form, Segment } from 'semantic-ui-react';
 import styled from 'styled-components';
 import { ICheck, ICheckData } from '../ICheck';
+import { groupIntoRows } from './utils';
 
 export interface CheckboxGroupOption {
   text: string;
@@ -59,22 +60,6 @@ const CheckboxGroup: FunctionComponent<ComponentProps> = (props) => {
     return props.selectedOptions && props.selectedOptions.length > 0
       ? !!props.selectedOptions.find((c: string) => c === value)
       : false;
-  };
-
-  const groupIntoRows = (options: CheckboxGroupOption[]): CheckboxGroupOption[][] => {
-    const rows: CheckboxGroupOption[][] = [];
-    const maxPerRow = 3;
-    for (let index = 0; index < options.length; index++) {
-      const option = options[index];
-      const latestRow = rows[rows.length - 1];
-      if (index % maxPerRow && latestRow?.length < maxPerRow) {
-        latestRow.push(option);
-      } else {
-        rows.push([option]);
-      }
-    }
-
-    return rows;
   };
 
   return (
